rename functions and moved into rdf module
- change the name of function cdf_s to avoid use the same name for the function and its output
- change name of function rdf_s
- change name of cdf_s
- move InterRDF_s into the rdf module
- add docs to explain difference InterRDF and InterRDF_s
- CHANGELOG and AUTHORS updated
