+++ package/AUTHORS
@@ -102,6 +102,8 @@ Chronological list of authors
   - Johannes Zeman
   - Ayush Suhane
   - Davide Cruz
+  - Shujie Fan
+
 
 External code
 -------------
@@ -150,3 +152,4 @@ Logo
 The MDAnalysis 'Atom' logo was designed by Christian Beckstein; it is
 Copyright (c) 2011 Christian Beckstein and made available under a
 Creative Commons Attribution-NoDerivs 3.0 Unported License.
+
