+++ package/CHANGELOG
@@ -13,7 +13,8 @@ The rules for this file:
   * release numbers follow "Semantic Versioning" http://semver.org
 
 ------------------------------------------------------------------------------
-??/??/18 tylerjereddy, richardjgowers, palnabarun, orbeckst, kain88-de, zemanj
+??/??/18 tylerjereddy, richardjgowers, palnabarun, orbeckst, kain88-de, zemanj,
+         VOD555
 
   * 0.18.1
 
@@ -21,6 +22,8 @@ Enhancements
   * Added various duecredit stubs
   * Import time reduced by a factor two (PR #1881)
   * added compound kwarg to center, centroid, center_of_geometry, center_of_mass (PR #1903)
+  * Added rdf.InterRDF_s to calculate site-specific pair distribution
+    functions (PR #1815)
 
 Fixes
   * Fixed order of indices in Angle/Dihedral/Improper repr
