+++ testsuite/MDAnalysisTests/analysis/test_rdf_s.py
@@ -2,7 +2,7 @@
 # vim: tabstop=4 expandtab shiftwidth=4 softtabstop=4 fileencoding=utf-8
 #
 # MDAnalysis --- https://www.mdanalysis.org
-# Copyright (c) 2006-2017 The MDAnalysis Development Team and contributors
+# Copyright (c) 2006-2018 The MDAnalysis Development Team and contributors
 # (see the file AUTHORS for the full list of names)
 #
 # Released under the GNU Public Licence, v2 or any higher version
@@ -27,7 +27,7 @@ import pytest
 from numpy.testing import assert_almost_equal
 
 import MDAnalysis as mda
-from MDAnalysis.analysis.rdf_s import InterRDF_s
+from MDAnalysis.analysis.rdf import InterRDF_s
 
 from MDAnalysisTests.datafiles import GRO_MEMPROT, XTC_MEMPROT
 
@@ -43,24 +43,24 @@ def sels(u):
     s2 = u.select_atoms('(name OD1 or name OD2) and resid 51 and sphzone 5.0 (resid 289)')
     s3 = u.select_atoms('name ZND and (resid 291 or resid 292)')
     s4 = u.select_atoms('(name OD1 or name OD2) and sphzone 5.0 (resid 291)')
-    ag = [[s1, s2], [s3, s4]]
-    return ag
+    ags = [[s1, s2], [s3, s4]]
+    return ags
 
 @pytest.fixture(scope='module')
 def rdf(u,sels):
     return InterRDF_s(u, sels).run()
 
 def test_nbins(u):
-    ag = sels(u)
-    rdf = InterRDF_s(u, ag, nbins=412).run()
+    ags = sels(u)
+    rdf = InterRDF_s(u, ags, nbins=412).run()
 
     assert len(rdf.bins) == 412
 
 
 def test_range(u):
-    ag = sels(u)
+    ags = sels(u)
     rmin, rmax = 1.0, 13.0
-    rdf = InterRDF_s(u, ag, range=(rmin, rmax)).run()
+    rdf = InterRDF_s(u, ags, range=(rmin, rmax)).run()
 
     assert rdf.edges[0] == rmin
     assert rdf.edges[-1] == rmax
@@ -69,7 +69,11 @@ def test_range(u):
 def test_count_size(rdf):
     # ZND vs OD1 & OD2
     # should see 2 elements in rdf.count
-    #
+    # 1 element in rdf.count[0]
+    # 2 elements in rdf.count[0][0]
+    # 2 elements in rdf.count[1]
+    # 2 elements in rdf.count[1][0]
+    # 2 elements in rdf.count[1][1]
     assert len(rdf.count) == 2
     assert len(rdf.count[0]) == 1
     assert len(rdf.count[0][0]) == 2
@@ -91,8 +95,8 @@ def test_double_run(rdf):
     assert len(rdf.count[1][1][0][rdf.count[1][1][0] == 3]) == 1
 
 def test_cdf(rdf):
-    rdf.cdf_s()
-    assert rdf.cdf_s[0][0][0][-1] == rdf.count[0][0][0].sum()/rdf.n_frames
+    rdf.get_cdf()
+    assert rdf.cdf[0][0][0][-1] == rdf.count[0][0][0].sum()/rdf.n_frames
 
 
 @pytest.mark.parametrize("density, value", [
@@ -101,4 +105,4 @@ def test_cdf(rdf):
 
 def test_density(u, sels, density, value):
     rdf = InterRDF_s(u, sels, density=density).run()
-    assert_almost_equal(max(rdf.rdf_s[0][0][0]), value)
+    assert_almost_equal(max(rdf.rdf[0][0][0]), value)
